+++ src/plonemeeting/restapi/serializer/annex.py
@@ -32,18 +32,18 @@ class BaseSerializeAnnexToJson(object):
     def _additional_values(self, result, additional_values):
         """Let include every values available from
            parent's categorized_elements."""
-
         ignored_values = ["allowedRolesAndUsers", "last_updated", "visible_for_groups"]
         parent = self.context.aq_parent
         infos = _categorized_elements(parent)[self.context.UID()]
         values = {
             k: v for k, v in infos.items()
             if k not in result and not (k.endswith("_url") or k in ignored_values)
-            and (additional_values == "*" or k in additional_values)
+            and ("*" in additional_values or k in additional_values)
         }
 
         result.update(values)
         return result
+
 
 @implementer(ISerializeToJson)
 @adapter(IAnnex, Interface)